+++ app/src/main/java/charstars/uscfit/Adapters/ActivityAdapter.java
@@ -42,7 +42,7 @@ public class ActivityAdapter extends RecyclerView.Adapter<ActivityAdapter.MyView
     public void onBindViewHolder(MyViewHolder holder, int position) {
         Activity activity = activitiesList.get(position);
         holder.category.setText(activity.getCategory());
-        holder.defaultCalorieValue.setText(Long.toString(activity.getDefaultCalorieValue()));
+        holder.defaultCalorieValue.setText(Double.toString(activity.getDefaultCalorieValue()));
     }
 
     @Override
